+++ core/mnh_funcs_system.pas
@@ -17,7 +17,7 @@ FUNCTION resetRandom_impl intFuncSignature;
     if not(context.checkSideEffects('resetRandom',tokenLocation,[se_alterContextState])) then exit(nil);
     result:=nil;
     if (params= nil) or  (params^.size=0) then begin context.getParent^.prng.resetSeed(0); result:=newVoidLiteral; end else
-    if (params<>nil) and (params^.size=1) and (arg0^.literalType=lt_int) then begin context.getParent^.prng.resetSeed(int0^.value); result:=newVoidLiteral; end;
+    if (params<>nil) and (params^.size=1) and (arg0^.literalType=lt_int) then begin context.getParent^.prng.resetSeed(int0^.value.toInt); result:=newVoidLiteral; end;
   end;
 
 FUNCTION random_imp intFuncSignature;
@@ -26,7 +26,7 @@ FUNCTION random_imp intFuncSignature;
     if not(context.checkSideEffects('random',tokenLocation,[se_alterContextState])) then exit(nil);
     if (params=nil) or (params^.size=0) then exit(newRealLiteral(context.getParent^.prng.realRandom))
     else if (params<>nil) and (params^.size=1) and (arg0^.literalType=lt_int) then begin
-      count:=int0^.value;
+      count:=int0^.value.toInt;
       if count>0 then begin
         result:=newListLiteral;
         for i:=1 to count do listResult^.appendReal(context.getParent^.prng.realRandom);
@@ -40,12 +40,12 @@ FUNCTION intRandom_imp intFuncSignature;
   VAR i,count:longint;
   begin
     if not(context.checkSideEffects('intRandom',tokenLocation,[se_alterContextState])) then exit(nil);
-    if (params<>nil) and (params^.size=1) and (arg0^.literalType=lt_int) then exit(newIntLiteral(context.getParent^.prng.intRandom(int0^.value)))
+    if (params<>nil) and (params^.size=1) and (arg0^.literalType=lt_int) then exit(newIntLiteral(context.getParent^.prng.intRandom(int0^.value.toInt)))
     else if (params<>nil) and (params^.size=2) and (arg0^.literalType=lt_int) and (arg1^.literalType=lt_int) then begin
-      count:=int1^.value;
+      count:=int1^.value.toInt;
       if count>=0 then begin
         result:=newListLiteral;
-        for i:=1 to count do listResult^.appendInt(context.getParent^.prng.intRandom(int0^.value));
+        for i:=1 to count do listResult^.appendInt(context.getParent^.prng.intRandom(int0^.value.toInt));
         exit(result);
       end;
     end;
@@ -71,8 +71,8 @@ FUNCTION beep_imp intFuncSignature;
     end else if (params<>nil) and (params^.size=2) and (arg0^.literalType=lt_int) and (arg1^.literalType=lt_int) then begin
       result:=newVoidLiteral;
       {$ifdef debugMode} writeln(stdErr,'        DEBUG: beep by function call'); {$endif}
-      windows.beep(int0^.value,
-                   int1^.value);
+      windows.beep(int0^.value.toInt,
+                   int1^.value.toInt);
     {$endif}
     end;
   end;
@@ -171,7 +171,7 @@ FUNCTION printTo_impl intFuncSignature;
 FUNCTION setExitCode_impl intFuncSignature;
   begin
     if (params<>nil) and (params^.size=1) and (arg0^.literalType=lt_int) and context.checkSideEffects('setExitCode',tokenLocation,[se_alterContextState]) then begin
-      ExitCode:=int0^.value;
+      ExitCode:=int0^.value.toInt;
       context.adapters^.setUserDefinedExitCode(ExitCode);
       result:=newVoidLiteral;
     end else result:=nil;
